Renamed screenshareActive to isScreenshareActive
